Move event to top level and rename to `ModifyCustomSpawnersEvent`
